fix(lint): fix 116 production clippy errors
- indexing_slicing: replace arr[i] with .get(i) + proper error handling
- mechanical lints: remove unnecessary casts, use strip_prefix, derive Default, etc.
- too_many_arguments: allow on 4 functions (design-level, out of scope to refactor)

No behavior changes; all tests pass.

+++ src/math/ops/round.rs
@@ -14,9 +14,9 @@ pub fn round(a: &CasNumber) -> CasNumber {
             let half = *d / 2;
             let abs_r = (abs_n + half) / *d;
             if *n < 0 {
-                CasNumber::Int(-(abs_r as i64))
+                CasNumber::Int(-abs_r)
             } else {
-                CasNumber::Int(abs_r as i64)
+                CasNumber::Int(abs_r)
             }
         }
         _ => CasNumber::Int(a.to_f64().round() as i64),
